+++ src/main/java/com/gg/gong9/global/exception/exceptions/auth/AuthException.java
@@ -1,7 +1,6 @@
 package com.gg.gong9.global.exception.exceptions.auth;
 
 import com.gg.gong9.global.exception.BaseException;
-import com.gg.gong9.global.exception.ExceptionMessage;
 
 public class AuthException extends BaseException {
     public AuthException(AuthExceptionMessage message) {
