+++ src/main/java/com/gg/gong9/global/exception/exceptions/product/ProductException.java
@@ -1,10 +1,9 @@
-package com.gg.gong9.global.exception.exceptions;
+package com.gg.gong9.global.exception.exceptions.product;
 
 import com.gg.gong9.global.exception.BaseException;
-import com.gg.gong9.global.exception.ExceptionMessage;
 
 public class ProductException extends BaseException {
-    public ProductException(ExceptionMessage message) {
+    public ProductException(ProductExceptionMessage message) {
         super(message.getText(), message.getStatus());
     }
 }
