+++ src/main/java/com/gg/gong9/global/exception/exceptions/product/ProductExceptionMessage.java
@@ -1,0 +1,18 @@
+package com.gg.gong9.global.exception.exceptions.product;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+
+@Getter
+@RequiredArgsConstructor
+public enum ProductExceptionMessage {
+    PRODUCT_NOT_FOUND("해당 상품이 존재하지 않습니다.", HttpStatus.NOT_FOUND),
+    PRODUCT_IMAGE_NOT_FOUND("해당 상품에 대한 이미지가 존재하지 않습니다.", HttpStatus.BAD_REQUEST);
+
+
+    private final String text;
+    private final HttpStatus status;
+}
+
+
