+++ src/main/java/com/gg/gong9/global/exception/exceptions/s3/S3Exception.java
@@ -1,10 +1,9 @@
-package com.gg.gong9.global.exception.exceptions;
+package com.gg.gong9.global.exception.exceptions.s3;
 
 import com.gg.gong9.global.exception.BaseException;
-import com.gg.gong9.global.exception.ExceptionMessage;
 
 public class S3Exception extends BaseException {
-    public S3Exception(ExceptionMessage message) {
+    public S3Exception(S3ExceptionMessage message) {
         super(message.getText(), message.getStatus());
     }
 }
