+++ src/main/java/com/gg/gong9/global/exception/exceptions/user/UserException.java
@@ -1,7 +1,6 @@
 package com.gg.gong9.global.exception.exceptions.user;
 
 import com.gg.gong9.global.exception.BaseException;
-import com.gg.gong9.global.exception.ExceptionMessage;
 
 public class UserException extends BaseException {
     public UserException(UserExceptionMessage message) {
